move helper methods to utils
